Add fonts-freefont-ttf to dependency list
The emulator compiles and runs cleanly on a minimal Debian 10 (Buster) desktop - the only thing which is missing is the FreeSans font which the emulator tries to load. I've confirmed that `fonts-freefont-ttf` package exists for both Debian 10, Ubuntu 18.04, and Ubuntu 20.04 so it seems safe to add it to the list of dependencies.